Fix issue with zeros
This commit fixed a few issues with the zero button where if the user
tried to type a small decimal such as 0.3 it would just show up as .3
which wasn't as nice. It also fixed a problem where the user could keep
adding zeros and make a long string of zeros with no other digits. The
ability to do this was removed because 000000 = 0.