Fix zeros bug
A bug occurred where if the user pressed a zero when all there was on
the display was a zero, the display would clear completely because
the numClick function removed the zero when it saw that the "number"
clicked wasn't a "." even when it was another zero. There was a line
checking if it was a zero but it came after and simply returned without
doing anything. The order was switched so that it would first check if
it was a zero, and if not then check if it was a "." and remove the
display content.